refactor(broker): rename package commandapi -> externalapi
Since this change moves the files to another folder and this can make
git forget about the files history, I've split this commit from the
class renaming
